service accept mkdir and rmdir as args
more flexible for user to choose use cases
fix: service should handle each directory using its own try block

+++ mosaic/utils/service.py
@@ -5,14 +5,30 @@ from mosaic.utils import ROOT_DIR, TEMP_DIR
 
 
 class service(ContextDecorator):
+    def __init__(
+        self,
+        *,
+        mkdir: bool = True,
+        rmdir: bool = True,
+    ) -> None:
+        super().__init__()
+        self.mkdir = mkdir
+        self.rmdir = rmdir
+
     def __enter__(self) -> Self:
-        ROOT_DIR.mkdir(exist_ok=True)
-        TEMP_DIR.mkdir(exist_ok=True)
+        if self.mkdir:
+            ROOT_DIR.mkdir(exist_ok=True)
+            TEMP_DIR.mkdir(exist_ok=True)
         return self
 
     def __exit__(self, *_) -> None:
-        try:
-            TEMP_DIR.rmdir()
-            ROOT_DIR.rmdir()
-        except OSError:
-            pass
+        if self.rmdir:
+            print('service: do rmdir')
+            try:
+                TEMP_DIR.rmdir()
+            except OSError:
+                pass
+            try:
+                ROOT_DIR.rmdir()
+            except OSError:
+                pass
